fully merged with zach
