+++ pom.xml
@@ -82,7 +82,6 @@
 			</exclusions>
 		</dependency>
 
-<<<<<<< HEAD
 		<!-- JMeter -->
 		<!-- https://mvnrepository.com/artifact/org.apache.jmeter/ApacheJMeter_http -->
 		<dependency>
@@ -102,7 +101,7 @@
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok</artifactId>
 			<scope>provided</scope>
-=======
+		</dependency>
 		<!-- https://mvnrepository.com/artifact/org.apache.jmeter/ApacheJMeter_core -->
 		<dependency>
 			<groupId>org.apache.jmeter</groupId>
@@ -136,7 +135,6 @@
 			<groupId>commons-logging</groupId>
 			<artifactId>commons-logging</artifactId>
 			<version>1.1.1</version>
->>>>>>> origin/JMeterService
 		</dependency>
 
 	</dependencies>
